IdentifyFlash returns u32 now

+++ agb_flash/agb_flash_1m.c
@@ -10,7 +10,7 @@ const struct FlashSetupInfo * const sSetupInfos[] =
     &DefaultFlash
 };
 
-u16 IdentifyFlash(void)
+u32 IdentifyFlash(void)
 {
     u16 result;
     u16 flashId;
